Trait polkadot_overseer_gen::MapSubsystem

pub trait MapSubsystem<T> {
    type Output;
    fn map_subsystem(&self, sub: T) -> Self::Output;
}

A helper trait to map a subsystem to smth. else.

Associated Types

type Output

The output type of the mapping.

Required methods

fn map_subsystem(&self, sub: T) -> Self::Output

Consumes a T per subsystem, and maps it to Self::Output.